Response factory refactor
