+++ src/Illuminate/Foundation/Console/Optimize/config.php
@@ -49,6 +49,7 @@ return array_map('realpath', array(
     $basePath.'/vendor/laravel/framework/src/Illuminate/Filesystem/FilesystemServiceProvider.php',
     $basePath.'/vendor/laravel/framework/src/Illuminate/Session/SessionServiceProvider.php',
     $basePath.'/vendor/laravel/framework/src/Illuminate/View/ViewServiceProvider.php',
+    $basePath.'/vendor/laravel/framework/src/Illuminate/Http/HttpServiceProvider.php',
     $basePath.'/vendor/laravel/framework/src/Illuminate/Routing/RouteFiltererInterface.php',
     $basePath.'/vendor/laravel/framework/src/Illuminate/Routing/Router.php',
     $basePath.'/vendor/laravel/framework/src/Illuminate/Routing/Route.php',
